--- licence.md
@@ -1,33 +1,0 @@
-USC-RL v1.0
-The Software is made available for academic or non-commercial purposes only. The license is for
-a copy of the program for an unlimited term. Individuals requesting a license for commercial use
-must pay for a commercial license.
-USC Stevens Institute for Innovation
-University of Southern California
-1150 S. Olive Street, Suite 2300
-Los Angeles, CA 90115, USA
-ATTN: Accounting
-DISCLAIMER. USC MAKES NO EXPRESS OR IMPLIED WARRANTIES, EITHER IN FACT OR BY
-OPERATION OF LAW, BY STATUTE OR OTHERWISE, AND USC SPECIFICALLY AND EXPRESSLY
-DISCLAIMS ANY EXPRESS OR IMPLIED WARRANTY OF MERCHANTABILITY OR FITNESS FOR A
-PARTICULAR PURPOSE, VALIDITY OF THE SOFTWARE OR ANY OTHER INTELLECTUAL PROPERTY
-RIGHTS OR NON-INFRINGEMENT OF THE INTELLECTUAL PROPERTY OR OTHER RIGHTS OF ANY
-THIRD PARTY. SOFTWARE IS MADE AVAILABLE AS-IS.
-LIMITATION OF LIABILITY. TO THE MAXIMUM EXTENT PERMITTED BY LAW, IN NO EVENT WILL
-USC BE LIABLE TO ANY USER OF THIS CODE FOR ANY INCIDENTAL, CONSEQUENTIAL, EXEMPLARY
-OR PUNITIVE DAMAGES OF ANY KIND, LOST GOODWILL, LOST PROFITS, LOST BUSINESS AND/OR
-ANY INDIRECT ECONOMIC DAMAGES WHATSOEVER, REGARDLESS OF WHETHER SUCH DAMAGES
-ARISE FROM CLAIMS BASED UPON CONTRACT, NEGLIGENCE, TORT (INCLUDING STRICT LIABILITY
-OR OTHER LEGAL THEORY), A BREACH OF ANY WARRANTY OR TERM OF THIS AGREEMENT, AND
-REGARDLESS OF WHETHER USC WAS ADVISED OR HAD REASON TO KNOW OF THE POSSIBILITY OF
-INCURRING SUCH DAMAGES IN ADVANCE.
-For commercial license pricing and annual commercial update and support pricing, please
-contact:
-<Licensing Associate Name>
-USC Stevens Institute for Innovation
-University of Southern California
-1150 S. Olive Street, Suite 2300
-Los Angeles, CA 90015, USA
-Tel: <Licensing Associate phone number>
-Fax: +1 213-821-5001
-Email: <Licensing Associate Email> and cc to: accounting@stevens.usc.edu
